next version; add a way to easily extend XNJS processing

+++ xnjs/src/main/java/eu/unicore/xnjs/ems/BasicManager.java
@@ -439,7 +439,7 @@ public class BasicManager implements Manager, InternalManager {
 	}
 
 	@Override
-	public void scheduleEvent(final XnjsEvent event, int time, TimeUnit units) throws RejectedExecutionException{
+	public void scheduleEvent(final XnjsEvent event, long time, TimeUnit units) throws RejectedExecutionException{
 		Runnable r=new Runnable(){
 			public void run(){
 				try{
